+++ packages/smooth_app/lib/cards/product_cards/smooth_product_base_card.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:smooth_app/generic_lib/widgets/smooth_card.dart';
+import 'package:smooth_app/widgets/smooth_card.dart';
 
 /// A common Widget for carrousel item cards.
 /// It allows to have the correct width/height and also a scale down feature,
