+++ packages/smooth_app/lib/cards/product_cards/smooth_product_card_found.dart
@@ -8,12 +8,12 @@ import 'package:smooth_app/data_models/product_preferences.dart';
 import 'package:smooth_app/data_models/user_preferences.dart';
 import 'package:smooth_app/generic_lib/design_constants.dart';
 import 'package:smooth_app/generic_lib/svg_icon_chip.dart';
-import 'package:smooth_app/generic_lib/widgets/smooth_card.dart';
-import 'package:smooth_app/generic_lib/widgets/smooth_product_image.dart';
 import 'package:smooth_app/helpers/product_cards_helper.dart';
 import 'package:smooth_app/helpers/product_compatibility_helper.dart';
 import 'package:smooth_app/helpers/ui_helpers.dart';
-import 'package:smooth_app/pages/product/new_product_page.dart';
+import 'package:smooth_app/pages/product_page/new_product_page.dart';
+import 'package:smooth_app/widgets/smooth_card.dart';
+import 'package:smooth_app/widgets/smooth_product_image.dart';
 
 class SmoothProductCardFound extends StatelessWidget {
   const SmoothProductCardFound({
