+++ packages/smooth_app/lib/pages/onboarding/country_selector.dart
@@ -5,8 +5,8 @@ import 'package:openfoodfacts/utils/CountryHelper.dart';
 import 'package:smooth_app/data_models/user_preferences.dart';
 import 'package:smooth_app/generic_lib/design_constants.dart';
 import 'package:smooth_app/generic_lib/dialogs/smooth_alert_dialog.dart';
-import 'package:smooth_app/generic_lib/widgets/smooth_text_form_field.dart';
 import 'package:smooth_app/query/product_query.dart';
+import 'package:smooth_app/widgets/smooth_text_form_field.dart';
 
 /// A selector for selecting user's country.
 class CountrySelector extends StatefulWidget {
