+++ packages/smooth_app/lib/pages/onboarding/preferences_page.dart
@@ -12,7 +12,7 @@ import 'package:smooth_app/generic_lib/design_constants.dart';
 import 'package:smooth_app/pages/onboarding/next_button.dart';
 import 'package:smooth_app/pages/onboarding/onboarding_flow_navigator.dart';
 import 'package:smooth_app/pages/preferences/user_preferences_food.dart';
-import 'package:smooth_app/pages/product/summary_card.dart';
+import 'package:smooth_app/pages/product_page/summary_card.dart';
 
 class PreferencesPage extends StatefulWidget {
   const PreferencesPage(this._localDatabase, this.backgroundColor) : super();
