+++ packages/smooth_app/lib/pages/product_page/summary_card.dart
