+++ packages/smooth_app/lib/pages/scan/scan_product_card.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:openfoodfacts/model/Product.dart';
 import 'package:provider/provider.dart';
 import 'package:smooth_app/data_models/product_preferences.dart';
-import 'package:smooth_app/pages/product/new_product_page.dart';
-import 'package:smooth_app/pages/product/summary_card.dart';
+import 'package:smooth_app/pages/product_page/new_product_page.dart';
+import 'package:smooth_app/pages/product_page/summary_card.dart';
 
 class ScanProductCard extends StatelessWidget {
   ScanProductCard(this.product)
