+++ packages/smooth_app/lib/widgets/images/smooth_image.dart
