+++ packages/smooth_app/lib/widgets/images/smooth_images_view.dart
