+++ packages/smooth_app/lib/widgets/picture_not_found.dart
