+++ packages/smooth_app/lib/widgets/smooth_card.dart
