+++ packages/smooth_app/lib/widgets/smooth_error_card.dart
