+++ packages/smooth_app/lib/widgets/smooth_gauge.dart
