+++ packages/smooth_app/lib/widgets/smooth_list_tile_card.dart
