+++ packages/smooth_app/lib/widgets/smooth_product_image.dart
