+++ packages/smooth_app/lib/widgets/smooth_responsive.dart
